refactor(cli): flatten credit command module
Move the Credit command and focused test to the neutral command root, rename the command-owned declarations, and update the lazy registry entry without changing public CLI or legacy token behavior.
